repo_results/run-20240308-125808_LLMatic-ZeroShot-CodeT-SelfRefine-Reflexion-ZeroShot_donnemartin-system-design-oop-url_shortener-donnemartin-system-design-oop-twitter-donnemartin-system-design-oop-whatsapp_0_20-runs_log/twitter/0/LLMatic/posting_content_management.py
class Post:
	def __init__(self, user, content, hashtags):
		self.user = user
		self.content = content
		self.hashtags = hashtags
		self.likes = 0
		self.retweets = 0
		self.replies = []

	def like(self, user):
		self.likes += 1
		self.user.create_notification('like', self)

	def retweet(self, user):
		self.retweets += 1
		self.user.create_notification('retweet', self)

	def reply(self, user, content):
		self.replies.append(content)
		self.user.create_notification('reply', self)


class PostingContentManagement:
	def __init__(self):
		self.posts = []

	def create_post(self, user, content, hashtags):
		self.posts.append(Post(user, content, hashtags))
