repo_results/run-20240308-125808_LLMatic-ZeroShot-CodeT-SelfRefine-Reflexion-ZeroShot_donnemartin-system-design-oop-url_shortener-donnemartin-system-design-oop-twitter-donnemartin-system-design-oop-whatsapp_0_20-runs_log/twitter/0/LLMatic/test_posting_content_management.py
import pytest
from posting_content_management import PostingContentManagement, Post
from user_management import User


def test_like_post():
	content_management = PostingContentManagement()
	user = User('email', 'username', 'password')
	post = Post(user, 'Hello!', [])
	content_management.posts.append(post)
	user.like(post)
	assert post.likes == 1


def test_retweet_post():
	content_management = PostingContentManagement()
	user = User('email', 'username', 'password')
	post = Post(user, 'Hello!', [])
	content_management.posts.append(post)
	user.retweet(post)
	assert post.retweets == 1


def test_reply_post():
	content_management = PostingContentManagement()
	user = User('email', 'username', 'password')
	post = Post(user, 'Hello!', [])
	content_management.posts.append(post)
	user.reply(post, 'Reply!')
	assert len(post.replies) == 1
